Changing floats to doubles.

+++ EnturDataConnector/src/main/java/fi/ahto/example/entur/data/connector/SiriDataPoller.java
@@ -208,10 +208,10 @@ public class SiriDataPoller {
         // vaf.setTransitType(decodeTransitType(mvh));
         vaf.setVehicleId(prefix + mvh.getVehicleRef().getValue());
         if (mvh.getBearing() != null) {
-            vaf.setBearing(mvh.getBearing().floatValue());
+            vaf.setBearing(mvh.getBearing().doubleValue());
         }
-        vaf.setLatitude(mvh.getVehicleLocation().getLatitude().floatValue());
-        vaf.setLongitude(mvh.getVehicleLocation().getLongitude().floatValue());
+        vaf.setLatitude(mvh.getVehicleLocation().getLatitude().doubleValue());
+        vaf.setLongitude(mvh.getVehicleLocation().getLongitude().doubleValue());
         if (mvh.getOriginAimedDepartureTime() != null) {
             vaf.setTripStart(mvh.getOriginAimedDepartureTime().withZoneSameInstant(ZoneId.of("Europe/Oslo")));
             vaf.setStartTime(GTFSLocalTime.ofZonedDateTime(mvh.getOriginAimedDepartureTime().withZoneSameInstant(ZoneId.of("Europe/Oslo"))));
