+++ HSLMQTTConnector/src/main/java/fi/ahto/example/hsl/data/mqtt/connector/HSLDataMQTTListener.java
@@ -153,8 +153,8 @@ public class HSLDataMQTTListener {
         vaf.setDelay(vp.path("dl").asInt());
         // Use Siri and GTFS-RT definition of the meaning of delay.
         vaf.setDelay(0 - vaf.getDelay());
-        vaf.setLatitude((float)vp.path("lat").asDouble());
-        vaf.setLongitude((float)vp.path("long").asDouble());
+        vaf.setLatitude(vp.path("lat").asDouble());
+        vaf.setLongitude(vp.path("long").asDouble());
         vaf.setDirection(vp.path("dir").asText());
         vaf.setRecordTime(Instant.ofEpochSecond(vp.path("tsi").asLong()));
 
@@ -174,8 +174,8 @@ public class HSLDataMQTTListener {
             vaf.setNextStopId(PREFIX + nextstop);
         }
         // New data available.
-        vaf.setBearing((float)vp.path("hdg").asDouble());
-        vaf.setSpeed((float)vp.path("spd").asDouble());
+        vaf.setBearing(vp.path("hdg").asDouble());
+        vaf.setSpeed(vp.path("spd").asDouble());
         return vaf;
     }
